% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bmt}
\alias{bmt}
\title{Bone merrow transplant data}
\format{A data frame with 166 observations and 6 variables:
\describe{
\item{agvhd}{Acute Graft Versus Host Disease }
\item{diag}{Diagnosis}
\item{agedon}{Age of donor}
\item{agerec}{Age of recipient}
\item{sexdon}{Sex of donor}
\item{sexrec}{Sex of recipient}
}}
\usage{
bmt
}
\description{
Bone merrow transplant data
}
\examples{
TODO
}
\keyword{datasets}
