% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fracture}
\alias{fracture}
\title{Fracture Data}
\format{A data frame with 36 observations and 5 variables:
\describe{
\item{ID}{}
\item{DEF}{}
\item{LOC}{}
\item{STEN}{}
}}
\usage{
fracture
}
\description{
Fracture Data
}
\examples{
TODO
}
\keyword{datasets}
