% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hosmerlemeshow.R
\name{hoslem.test}
\alias{hoslem.test}
\title{Hosmer lemeshaw goodness of fit test}
\usage{
hoslem.test(x, y, g = 10)
}
\arguments{
\item{Observed}{outcomes}

\item{Estimated}{probabilities}

\item{number}{of intervals}
}
\value{
A list with the test statistic and p value of the test
}
\description{
Hosmer lemeshaw goodness of fit test
}
\examples{
TODO
}
