% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LBW}
\alias{LBW}
\alias{LBW}
\title{Low Birth Weight Data}
\format{A data frame with 36 observations and 5 variables:
\describe{
\item{SMOKING}{Smoking}
\item{ALCOHOL}{Alcohol use}
\item{SOCCLASS}{Social Economic Class}
\item{LWB}{Low Birth Weight}
\item{NUMBERS}{Bumber of observations in groups}
}}
\usage{
LBW

LBW
}
\description{
Low Birth Weight Data

Fracture Data
}
\examples{
TODO
TODO
}
\keyword{datasets}
