% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ova}
\alias{ova}
\title{Ovarian cancer data}
\format{A data frame with 358 observations and 8 variables:
\describe{
\item{T}{}
\item{D}{}
\item{SURV}{}
\item{KARN}{}
\item{BRODERS}{}
\item{FIGO}{}
\item{ASCITES}{}
\item{DIAM}{}
}}
\usage{
ova
}
\description{
Ovarian cancer data
}
\examples{
TODO
}
\keyword{datasets}
