% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twobytwo.R
\name{twobytwo}
\alias{twobytwo}
\alias{twobytwo_or}
\alias{twobytwo_rd}
\alias{twobytwo_rr}
\title{Various risk measures for the 2x2 table}
\usage{
twobytwo_or(a, b, c, d, log = TRUE)

twobytwo_rd(a, b, c, d)

twobytwo_rr(a, b, c, d, log = TRUE)
}
\arguments{
\item{a}{number of deseased in the exposed group or a 2x2 matrix containing a, b, c and d}

\item{b}{number of non-deseased in the exposed group}

\item{c}{number of deseased in the non-exposed group}

\item{d}{number of non-deseased in the non-exposed group}

\item{log}{give results on the log scale (for the relative risk and the ods ratio the se is
not given on the original scale. For the risk difference the results are always given on
the original scale)}
}
\value{
A list of class 'riskestimate' with elements:
estimate: the (log)risk estimate
se: the standard error of the the (log)risk estimate
outcome: A description of the outcome
log: a logical value indicating if the estimate is on the log scale
}
\description{
Various risk measures for the 2x2 table
}
\examples{
\\cite{Van Linschoten et al. (2009)}
twobytwo_or(26, 62, 21, 60)
twobytwo_rr(26, 62, 21, 60)
twobytwo_rd(26, 62, 21, 60)
}
